refactor: Move onwards components to the root
We're doing this so that our dynamic import solution works which needs files to exist at the same level
