+++ dotcom-rendering/src/web/components/Carousel.stories.tsx
