+++ dotcom-rendering/src/web/components/Carousel.tsx
