+++ dotcom-rendering/src/web/components/Spotlight.tsx
